+++ libraries/kotlinx-metadata/jvm/ChangeLog.md
@@ -5,6 +5,7 @@
 - [`KT-31308`](https://youtrack.jetbrains.com/issue/KT-31308) Add module name extensions to kotlinx-metadata-jvm
 - [`KT-31338`](https://youtrack.jetbrains.com/issue/KT-31338) Retain "is moved from interface companion" property flag in kotlinx-metadata-jvm
     - Breaking change: JvmPropertyExtensionVisitor.visit has a new parameter `jvmFlags: Flags`
+- Correctly write "null" constant value in effect expression of a contract
 - Rename `desc` parameters to `signature` in JvmFunctionExtensionVisitor, JvmPropertyExtensionVisitor, JvmConstructorExtensionVisitor
 
 ## 0.0.5